+++ pila/stack.go
@@ -68,9 +68,22 @@ func (s *Stack) Push(element interface{}) {
 }
 
 // Pop removes and returns the element on top of the Stack.
-// If the Stack was empty, it returns false.
+// If the Stack is empty, it returns false.
 func (s *Stack) Pop() (interface{}, bool) {
 	return s.base.Pop()
+}
+
+// Sweep removes and returns the bottommost element of the Stack.
+// If the Stack is empty, it returns false.
+func (s *Stack) Sweep() (interface{}, bool) {
+	return s.base.Sweep()
+}
+
+// SweepPush removes and returns the bottommost element of the Stack,
+// and pushes an element on top of it, as an atomic operation.
+// If the Stack is empty, it returns false.
+func (s *Stack) SweepPush(element interface{}) (interface{}, bool) {
+	return s.base.SweepPush(element)
 }
 
 // Size returns the size of the Stack.